+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - `BetaCreateCreditCardClientSecret`
   - `BetaCreateBankAccountClientSecret`
   - `BetaCreateBankAccountClientSecretWithReturlUrl`
+- Routes `AmazonShippingAccount` to correct endpoint on create
 - Adds `TrackingCodes` to tracker list function
 - Fixes error parsing
   - Allows for alternative format of `errors` field (previously we deserialized the `errors` field into a list of `Error` objects; however, sometimes the errors are simply a list of strings. This change makes the `errors` field an `interface`, allowing for either the renamed `FieldError` object or a list of strings. Users will need to check for the type of error returned and handle appropriately)